requirements.txt
asgiref==3.8.1
crispy-bootstrap5==2024.2
Django==5.0.4
django-crispy-forms==2.1
Faker==25.2.0
gunicorn==22.0.0
joblib==1.4.2
numpy==1.26.4
packaging==24.0
pandas==2.2.2
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-ulid==2.5.0
pytz==2024.1
scikit-learn==1.5.0
scipy==1.13.0
six==1.16.0
sqlparse==0.4.4
threadpoolctl==3.5.0
typing_extensions==4.11.0
tzdata==2024.1
whitenoise==6.6.0
